renamed BgListPrice -> TradeDeck
